+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [2.2.0](https://github.com/netlify/primitives/compare/vite-plugin-v2.1.4...vite-plugin-v2.2.0) (2025-06-06)
+
+
+### Features
+
+* image cdn support ([#232](https://github.com/netlify/primitives/issues/232)) ([01c844d](https://github.com/netlify/primitives/commit/01c844d82a27a9812be7634219d9bdc69a128985))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.1.4 to 4.2.0
+    * @netlify/dev-utils bumped from ^3.1.1 to ^3.2.0
+
+## [2.1.4](https://github.com/netlify/primitives/compare/vite-plugin-v2.1.3...vite-plugin-v2.1.4) (2025-06-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @netlify/dev bumped from 4.1.3 to 4.1.4
+
 ## [2.1.3](https://github.com/netlify/primitives/compare/vite-plugin-v2.1.2...vite-plugin-v2.1.3) (2025-06-04)